Fix several cases of undefined behaviour, memory corruption and memory leaks

+++ core/src/lib/bsys.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -556,7 +556,7 @@ void WritePidFile(int pidfile_fd,
 int DeletePidFile(const char* pidfile_path)
 {
 #if !defined(HAVE_WIN32)
-  unlink(pidfile_path);
+  if (pidfile_path) { unlink(pidfile_path); }
 #endif
   return 1;
 }
